Componentize connection handshake components
