+++ crates/relayer-components/src/chain/traits/components/connection_handshake_message_builder.rs
@@ -1,45 +1,12 @@
-use cgp_core::{async_trait, HasErrorType};
+use cgp_core::prelude::*;
+use cgp_core::HasErrorType;
 
-use crate::chain::traits::types::client_state::HasClientStateType;
 use crate::chain::traits::types::connection::{
     HasConnectionHandshakePayloads, HasInitConnectionOptionsType,
 };
 use crate::std_prelude::*;
 
-#[async_trait]
-pub trait CanBuildConnectionHandshakePayloads<Counterparty>:
-    HasConnectionHandshakePayloads<Counterparty> + HasClientStateType<Counterparty> + HasErrorType
-{
-    async fn build_connection_open_init_payload(
-        &self,
-        client_state: &Self::ClientState,
-    ) -> Result<Self::ConnectionOpenInitPayload, Self::Error>;
-
-    async fn build_connection_open_try_payload(
-        &self,
-        client_state: &Self::ClientState,
-        height: &Self::Height,
-        client_id: &Self::ClientId,
-        connection_id: &Self::ConnectionId,
-    ) -> Result<Self::ConnectionOpenTryPayload, Self::Error>;
-
-    async fn build_connection_open_ack_payload(
-        &self,
-        client_state: &Self::ClientState,
-        height: &Self::Height,
-        client_id: &Self::ClientId,
-        connection_id: &Self::ConnectionId,
-    ) -> Result<Self::ConnectionOpenAckPayload, Self::Error>;
-
-    async fn build_connection_open_confirm_payload(
-        &self,
-        client_state: &Self::ClientState,
-        height: &Self::Height,
-        client_id: &Self::ClientId,
-        connection_id: &Self::ConnectionId,
-    ) -> Result<Self::ConnectionOpenConfirmPayload, Self::Error>;
-}
-
+#[derive_component(ConnectionHandshakeMessageBuilderComponent, ConnectionHandshakeMessageBuilder<Chain>)]
 #[async_trait]
 pub trait CanBuildConnectionHandshakeMessages<Counterparty>:
     HasInitConnectionOptionsType<Counterparty> + HasErrorType
