+++ crates/relayer-components/src/chain/traits/message_builders/mod.rs
@@ -1,4 +1,3 @@
-pub mod connection;
 pub mod create_client;
 pub mod receive_packet;
 pub mod timeout_unordered_packet;
