+++ crates/relayer-components/src/relay/impls/connection/open_confirm.rs
@@ -2,9 +2,8 @@ use cgp_core::async_trait;
 
 use crate::chain::traits::client::client_state::CanQueryClientState;
 use crate::chain::traits::components::chain_status_querier::CanQueryChainHeight;
-use crate::chain::traits::message_builders::connection::{
-    CanBuildConnectionHandshakeMessages, CanBuildConnectionHandshakePayloads,
-};
+use crate::chain::traits::components::connection_handshake_message_builder::CanBuildConnectionHandshakeMessages;
+use crate::chain::traits::components::connection_handshake_payload_builder::CanBuildConnectionHandshakePayloads;
 use crate::relay::traits::chains::HasRelayChains;
 use crate::relay::traits::components::ibc_message_sender::{CanSendSingleIbcMessage, MainSink};
 use crate::relay::traits::components::update_client_message_builder::CanBuildUpdateClientMessage;
