+++ crates/relayer-components/src/relay/impls/connection/open_try.rs
@@ -4,9 +4,8 @@ use cgp_core::async_trait;
 
 use crate::chain::traits::client::client_state::CanQueryClientState;
 use crate::chain::traits::components::chain_status_querier::CanQueryChainHeight;
-use crate::chain::traits::message_builders::connection::{
-    CanBuildConnectionHandshakeMessages, CanBuildConnectionHandshakePayloads,
-};
+use crate::chain::traits::components::connection_handshake_message_builder::CanBuildConnectionHandshakeMessages;
+use crate::chain::traits::components::connection_handshake_payload_builder::CanBuildConnectionHandshakePayloads;
 use crate::chain::traits::types::ibc::HasIbcChainTypes;
 use crate::chain::traits::types::ibc_events::connection::HasConnectionOpenTryEvent;
 use crate::relay::traits::chains::HasRelayChains;
